Create logic for new grid button
First, I moved the new grid button to the top of the page to match the
project specifications. I added a prompt to the new grid button to ask
the use for how large the grid should be. The input is checked to
make sure the number is less than 100 and is a number. If the requested
number is over 100 or not actually a number then, the user is given
the prompt again. When a correct input is receive, the old grid is
cleared and a new grid with the requested number of rows and columns
is created.