Changed hover logic
Up until this point, when the mouse enter a box, the background
would change to a light green color, and upon the mouse leaving the box,
the color would be reset to the background color. When clicked on, the
background would be set to a dark green color at 10% opacity, and the
entering and leaving color change would be disabled. Each subsequent
click would add another 10% to the opacity.

The exiting logic was removed, so that the mouse now leaves a trail. The
opacity change was also moved to the mouseenter check so now each
mouseover adds 10% to the opacity.

The original assignment wanted it this way, and I thought that the
clicking to set the color would be neater. I was wrong, it was annoying
to do, so I reverted to the assignment specs.